fix: split hook for testing to work
